// frontend-ui/src/app/services/models/installment-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Agent } from '../models/agent';
import { Member } from '../models/member';
import { PaymentSchedule } from '../models/payment-schedule';
import { Product } from '../models/product';

/**
 * Installment response with all details
 */
export interface InstallmentResponseDto {

  /**
   * Advance payment
   */
  advanced_paid: number;

  /**
   * Creation timestamp
   */
  createdTime?: string;
  given_product_agent: Agent;

  /**
   * Unique identifier
   */
  id?: number;

  /**
   * Image file paths
   */
  imageFilePaths?: Array<string>;

  /**
   * Number of installment months
   */
  installmentMonths: number;

  /**
   * Interest rate percentage
   */
  interestRate: number;
  member: Member;

  /**
   * Monthly payment amount
   */
  monthlyInstallmentAmount?: number;

  /**
   * Amount remaining to be paid
   */
  needPaidAmount?: number;

  /**
   * Additional costs
   */
  otherCost: number;

  /**
   * Payment schedules
   */
  paymentSchedules?: Array<PaymentSchedule>;
  product: Product;

  /**
   * Installment status
   */
  status: 'PENDING' | 'ACTIVE' | 'COMPLETED' | 'OVERDUE' | 'CANCELLED' | 'DEFAULTED';

  /**
   * Total product amount
   */
  totalAmountOfProduct: number;

  /**
   * Total amount with interest
   */
  totalAmountWithInterest?: number;
}